fix: patch onboarding profile and notifications
