+++ App/shared/tokenLogger.ts
@@ -1,0 +1,6 @@
+import { useAuthStore } from '@/state/authStore';
+
+export function logTokenIssue(context: string) {
+  const { uid } = useAuthStore.getState();
+  console.warn(`🔐 Token issue during ${context}`, { uid });
+}
